refactor NewDialog to WelcomeDialog
