+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { Outlet } from "react-router-dom";
 import NavBarNew from "./components/nav-new/NavBarNew.jsx";
-import NewDialog from "./components/nav-new/NewDialog.jsx";
 import TutorialDialog from "./components/nav-new/TutorialDialog.jsx";
+import WelcomeDialog from "./components/nav-new/WelcomeDialog.jsx";
 
 function App() {
   const [isWelcome, setWelcome] = useState(
@@ -24,7 +24,7 @@ function App() {
         isTutorial={isTutorial}
         toggleTutorial={() => setTutorial(!isTutorial)}
       />
-      <NewDialog
+      <WelcomeDialog
         isWelcome={isWelcome}
         toggleWelcome={() => setWelcome(!isWelcome)}
       />
